% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_kf_helpers.R
\name{data_gen_lgssm}
\alias{data_gen_lgssm}
\title{Generates simulated data from a multivariate LGSSM.}
\usage{
data_gen_lgssm(TT, dim_x = 1, dim_y = 1, A, B, C, D, Q, R, initX)
}
\arguments{
\item{TT}{integer giving the time series length}

\item{dim_x}{integer giving the state process dimension; defaults to
\code{dim_x = 1}}

\item{dim_y}{integer giving the measurement process dimension; defaults to
\code{dim_x = 1}}

\item{A}{state process matrix1}

\item{B}{state process matrix2}

\item{C}{measurement process matrix1}

\item{D}{measurement process matrix2}

\item{Q}{VCM state process}

\item{R}{VCM measurement process}

\item{initX}{Initial value of state process (must have same length as
\code{dim_x}); can be thought of X_{t=0}, where time series runs from
\code{t=1,...,TT}}
}
\value{
a named list of 7:
  \itemize{
    \item matrix or vector of simulated state process: \code{dim_x x TT}
    \item matrix or vector of simulated measurements: \code{dim_y x TT}
    \item matrix or vector of simulated regressors for states:
      \code{ncol(B) x TT}
    \item matrix or vector of simulated regressors for measurements:
      \code{ncol(D) x TT}
    \item initial state regressor values: \code{ncol(B) x 1}
    \item initial measurement regressor values: \code{ncol(D) x 1}
    \item \code{initX}, see argument to function call
  }
}
\description{
Generates simulated data from a multivariate LGSSM.
}
