% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBMdata.R
\docType{data}
\name{IBMdata}
\alias{IBMdata}
\title{Daily IBM log-returns.}
\format{
A data frame with 3271 rows and 1 variable:
\describe{
  \item{r_t}{daily log-return}
}
}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
data(IBMdata)
}
\description{
A dataset containing the daily IBM log-returns \eqn{r_t=log(s_t/s_{t-1})},
where \eqn{s_t} denotes the stock price (closing to closing) at day \eqn{t}
from Jan. 3rd, 2006 - Dec. 31, 2018 with overall series length \eqn{T=3271}.
}
\keyword{datasets}
