% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_kf_JS.R
\name{kfJSD}
\alias{kfJSD}
\title{Kalman joint smoothing}
\usage{
kfJSD(uReg, xtt, Ptt, A, B, Q, dimX, TT, nSimJS = 1)
}
\arguments{
\item{uReg}{Matrix (vector) of regressors for the latent state process of
dimension \code{ncol(B) x T}. For a single regressors \code{uReg} is a
vector of length \code{T}.}

\item{xtt}{forward filtering means as produced by \code{\link{kfMFPD}}}

\item{Ptt}{forward filtering variances as produced by \code{\link{kfMFPD}}}

\item{A}{Parameter (or system) matrix of dimension \code{dimX x dimX}.}

\item{B}{Parameter (or system) matrix of dimension \code{dimX x numU}.}

\item{Q}{Error VCM of state process of dimension \code{dimX x dimX}.}

\item{dimX}{integer giving the dimension of the latent state process}

\item{TT}{integer giving the length of the time series}

\item{nSimJS}{Number of joint smoothing (backward simulation) runs; defaults
to \code{nSimJS=1}.}
}
\value{
a named list of three:
  \itemize{
  \item\code{jsdEXP:} an array of dimension \code{dimX x TT x nSimJS} giving
     the smoothing means \eqn{\mu_t} of dimension \code{dimX} for all
     \eqn{t=TT,\ldots,1} per \eqn{1,\ldots,}\code{nSimJS} simulation run
  \item\code{jsdVAR:} an array of dimension \code{dimX x dimX  TT} giving
     the smoothing variances \eqn{L_t} of dimension \code{dimX x dimX} for
     all \eqn{t=TT,\ldots,1} (note: these do not change with the number of
     simulation runs \code{nSimJS}, in contrast to previous means)
  \item\code{jsdTRJ:} an array of dimension \code{dimX x TT x nSimJS} giving
     the sampled trajectories \eqn{\tilde{x}_{1:T}} (\code{dimX x TT}) from
     \eqn{p(x_{1:T}|y_{1:T})} for each simulation \code{nSimJS}
  }
}
\description{
Performs joint smoothiong via backward simulation i.e. generating
\code{nSimJS} realizations from \eqn{p(x_{1:T}|y_{1:T})}, as described in
the corresponding \emph{"Joint smoothing density- implemented backward
simulation recursions"} subsection of the \code{Details} section in
\code{\link{kfLGSSM}}.
}
