% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_kf_MF.R
\name{kfMFPD}
\alias{kfMFPD}
\title{Kalman marginal (forward) filtering computation and prediction}
\usage{
kfMFPD(
  yObs,
  uReg,
  wReg,
  dimX,
  dimY,
  TT,
  x00,
  P00,
  u00,
  A,
  B,
  C,
  D,
  Q,
  R,
  PDSTORE = FALSE
)
}
\arguments{
\item{yObs}{A matrix or vector of measurements (observations):
 \itemize{
 \item{rows: multivariate dimension}
 \item{columns: time series dimension \code{T}}
 }
If \code{Y} is a univariate process, \code{yObs} can be passed as a vector
of length \code{T}. If \code{nrow(yObs) = 1}, then \code{yObs} becomes a
vector of length \code{T}.}

\item{uReg}{Matrix (vector) of regressors for the latent state process of
dimension \code{ncol(B) x T}. For a single regressors \code{uReg} is a
vector of length \code{T}.}

\item{wReg}{Matrix (vector) of regressors for the measurement process of
dimension \code{ncol(D) x T}. For a single regressors \code{wReg} is a
vector of length \code{T}.}

\item{dimX}{integer giving the dimension of the latent state process}

\item{dimY}{integer giving the dimension of the measurement process}

\item{TT}{integer giving the length of the time series}

\item{x00}{see \code{initX} from \code{\link{kfLGSSM}}}

\item{P00}{see \code{initP} from \code{\link{kfLGSSM}}}

\item{u00}{see \code{initU} from \code{\link{kfLGSSM}}}

\item{A}{Parameter (or system) matrix of dimension \code{dimX x dimX}.}

\item{B}{Parameter (or system) matrix of dimension \code{dimX x numU}.}

\item{C}{Parameter (or system) matrix of dimension \code{dimY x dimX}.}

\item{D}{Parameter (or system) matrix of dimension \code{dimY x numW}.}

\item{Q}{Error VCM of state process of dimension \code{dimX x dimX}.}

\item{R}{Error VCM of measurement process of dimension \code{dimY x dimY}}

\item{PDSTORE}{logical; if \code{TRUE} prediction quantities
\eqn{\hat{P}_{t|t-1}} and \eqn{\hat{x}_{t|t-1}} are stored
\eqn{\forall t=1,\ldots,T}; otherwise, only stored for current
iteration \eqn{t} and not returned.}
}
\value{
if \code{PDSTORE = FALSE} a named list of two
  \itemize{
  \item \code{mfdEXP} a matrix of dimension \code{dimX x TT} with each
     column being the corresponding \eqn{\hat{x}_{t|t}} (see the
     \code{Details} section in \code{\link{kfLGSSM}}).
  \item \code{mfdVAR} an array of dimension \code{dimX x dimX x TT} with
  matrices \eqn{\hat{P}_{t|t}} of dimension \code{dimX x dimX} \eqn{\forall
  t = 1,\ldots,TT} (see the \code{Details} section in \code{\link{kfLGSSM}}).
  }
  if \code{PDSTORE = TRUE} a named list of four:
  \itemize{
  \item \code{mfdEXP} as above
  \item \code{mfdVAR} as above
  \item \code{pddEXP} a matrix of dimension \code{dimX x (TT + 1)} with each
     column being the corresponding \eqn{\hat{x}_{t+1|t}}, starting from
     \eqn{\hat{x}_{1|0}} and running to \eqn{\hat{x}_{T+1|T}} (see the
     \code{Details} section in \code{\link{kfLGSSM}}).
  \item \code{pddVAR} an array of dimension \code{dimX x dimX x (TT + 1)}
  with matrices \eqn{\hat{P}_{t+1|t}} of dimension \code{dimX x dimX},
  \eqn{\forall t = 1, \ldots,TT + 1}, starting with \eqn{\hat{P}_{1|0}} and
  running to \eqn{\hat{P}_{T+1|T}} (see the \code{Details} section in
  \code{\link{kfLGSSM}}).
  }
}
\description{
Runs (Marginal) filtering and prediction recursions, as described in the
corresponding \emph{"(Marginal) Filtering and prediction - implemented
recursions"} subsection of the \code{Details} section in
\code{\link{kfLGSSM}}.
}
