% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_kf_MS.R
\name{kfMSD}
\alias{kfMSD}
\title{Kalman marginal smoothing}
\usage{
kfMSD(dimX, TT, xtt, Ptt, xtt1, Ptt1, A, Q)
}
\arguments{
\item{dimX}{integer giving the dimension of the latent state process}

\item{TT}{integer giving the length of the time series}

\item{xtt}{forward filtering means as produced by \code{\link{kfMFPD}}}

\item{Ptt}{forward filtering variances as produced by \code{\link{kfMFPD}}}

\item{xtt1}{predictive means as produced by \code{\link{kfMFPD}} (if
\code{PDSTORE = TRUE})}

\item{Ptt1}{predictive variances as produced by \code{\link{kfMFPD}} (if
\code{PDSTORE = TRUE})}

\item{A}{Parameter (or system) matrix of dimension \code{dimX x dimX}.}

\item{Q}{Error VCM of state process of dimension \code{dimX x dimX}.}
}
\value{
a named list of two:
  \itemize{
  \item\code{msdEXP:} predicitve means \eqn{x_{t+1|t}}
     (see \emph{"Marginal smoothing - implemented recursions"} in
     \code{\link{kfLGSSM}})
  \item\code{msdEXP:} predicitve variances \eqn{P_{t+1|t}}
     (see \emph{"Marginal smoothing - implemented recursions"} in
     \code{\link{kfLGSSM}})
  }
}
\description{
Performs marginal smoothiong i.e. computation of means and variances for
\eqn{p(x_t|y_{1:T})\forall t=1,\ldots,TT}, as described in the corresponding
\emph{"Marginal smoothing - implemented recursions"} subsection of the
\code{Details} section in \code{\link{kfLGSSM}}.
}
