% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppSMCkalman.R
\docType{package}
\name{RcppSMCkalman}
\alias{RcppSMCkalman}
\title{Implementing multivariate Kalman filtering and smoothing with RcppSMC.}
\description{
Missing package description
}
\section{Missing section.}{
NA
}

