\name{RcppSMCkalmanFilter-package}
\alias{RcppSMCkalmanFilter-package}
\alias{RcppSMCkalmanFilter}
\docType{package}
\title{\packageTitle{RcppSMCkalmanFilter}}
\description{\packageDescription{RcppSMCkalmanFilter}}
\section{Package Content}{\packageIndices{RcppSMCkalmanFilter}}
\author{\packageAuthor{RcppSMCkalmanFilter}}
\section{Maintainer}{\packageMaintainer{RcppSMCkalmanFilter}}
\keyword{package}
